Clean up the threading model for wt bindings.
Connections are refcounted by the sessions it creates. This avoids passing around
the lifetime of the Connection as a viral lifetime attribute, but still has the
same features in terms of preventing a (consuming) close unless we can obtain
ownership from the Arc. This pattern is not repeated for Session & RecordCursor
as cursors are expected to be shorter lived.

Mark both Session and RecordCursor as Send -- according to docs they can be freely
used by threads other than the creating thread. Do not mark them as Sync as they
should not be used concurrently from multiple threads without synchronization.
In cases where they are shared a Mutex would allow it to be Sync (RwLock would not).