Implement Vamana search algorithm
Add abstractions for a DiskANN Graph and navigational vector store as
well as scoring.

Implement Vamana graph search in terms of these primitives. This is a
required step before
building bulk graph indexing as generating the candidate edge list on
node insertion starts
with a search of the partially built graph.