Encode metadata in the graph table (#6)
Encode graph metadata as json and write it into a special key in the
table so that it can be read
at startup. The content of this key is intended to be immutable.

In `bulk_load` compute an entry point and write this to another special
key. WT searchers will
then read this key later before they begin the search. This value should
not be cached as it could
change over time.

Fix a couple of issues:
* Restrict searcher during indexing so that edges to self are not
considered.
* Run search inside of a transaction during bulk build, otherwise each
individual read is considered a transaction which has considerable
costs.