Add search to `et` tool (#7)
Reads an input set of numpy vectors and runs them one at a time through
a graph search.
* Encode graph metadata and entry point in the graph table so we can
read it back later.
* Fix a bug in WiredTigerGraphNode that would return erroneously sized
vectors if the value memory was aligned.

Initial run with a 2GB cache yields 16ms/search which is not terribly
competitive on overall latency but likely
better than lucene in a small container. 75% of the profile is in
`WiredTigerGraph.get() -> pread()` which is
not unexpected and could certainly be improved.