Remove lifetime from `RecordCursor` (#9)
`RecordCursor` lifetimes make it more difficult to build other templated
abstractions.

In order to make this work we have a reference counted internal session
that is distinct from `Session`.
Each `RecordCursor` takes a reference to the inner session to ensure
that the underlying `WT_SESSION`
remains alive so long as there are any open cursors. We need to use
`Arc` to support sending `Session`
across threads, which also means that the inner session must be `Send +
Sync` which also causes
`Session` to be `Send + Sync`. Make most `Session` methods `&mut` which
will prevent use from
multiple threads anyway.

One caveat to this approach is that it is no longer possible to access
the `Session` from a `RecordCursor`
since that could be unsound. If this is important we could refactor the
representation so that thread-safe
`Session` access is possible via internal `Mutex`.

Remove most public `close()` functions because I'm pretty sure they will
leak memory if an error occurs.