Implement a `Session`-level cache of `RecordCursor` objects (#14)
This part 1 of N to implement intra-query concurrency in graph search.

Ultimately I'd like to reintroduce the lifetime parameter on
`RecordCursor`, in part because this makes it easier to
navigate the object hierarchy (back to `Session`) and in part because it
makes it harder to `Send` a `Session` with
open cursors. Caching cursors in the `Session` makes the rust-perceived
lifetime of the cursor shorter, which is good
for these purposes, and will make it easier to flatten our graph
traversal abstractions and remove associated types.

This changes `Session` and `RecordCursor` so that they are no longer
`Send` or `Sync`. As currently factored
sending these objects is unsound as I cannot perform perform two reads
concurrently on different threads -- this
mutates state in the `Session` and causes UB.

Move `RecordCursor` into the `session` module since they have a number
of private interfaces with each other.