Attach RecordCursor lifetime to underlying Session (#15)
This is part 2 of 3 to implement intra-query in graph search.

Mark `Session` as `Send` -- in practice this can only be used if there
are no outstanding cursors, although this
is less of an impediment given the new cursor caching mechanisms.

Fix various graph related traits to accept lifetimes, which is the
reason that `RecordCursor` lifetimes were removed
in the first place.

Rename uses of 'node' to 'vertex' to minimize confusion as node is
heavily overloaded.

Part 3 will provide a separate trait for concurrent lookups, a WT
implementation of that trait, and a concurrent
search implementation.