Implement concurrent graph vertex lookups in search (#16)
Provide a pool of threads and a concurrency limit and the searcher will
perform up to N graph lookups
on different threads, which reduces latency by 60-70% in isolated tests.

Introduce a `WorkerPool` concept where each thread owns a WT `Session`
that is used to perform the
lookups. A `lookup()` method is added to the graph abstraction passed to
the searcher.

In the algorithm itself we keep track of the number of in-flight reads
and each iteration through the loop
pop off enough candidates to keep the read queue saturated. When reads
are complete the vertex contents
is written to an mpsc channel that the driving thread consumes. This
minimizes locking and keeps the
threads focused on IO operations rather than performing any CPU
intensive work.

This is a capstone of several PRs that fixed underlying issues with the
WT model in rust.