During bulk load write a temp file containing nav vectors (#21)
The temp file can be mmapped and traversed like the input vectors,
eliminating the 50% of
CPU time we spend in wt cursor search. This should be a short term
productivity improvement
but more needs to be done to improve search costs since we seem to read
nav vectors
proportional to `max_edges * num_candidates`.