Add call to restrict `RecordCursor` bounds (#24)
`set_bounds` accepts a `RangeBound` to restrict positioning. This can be
useful if we want to iterate over a subset
of the input table -- we can set bounds and then use the cursor as an
`Iterator`.

Add a `wt_call!` macro to try to make programming in `wt_mdb` more
pleasant.