Remove intra-query concurrency implementation (#40)
This option allowed using up to N threads to perform reads within a
query.

Aside from the complexity, there are several issues with this:
* Inter-query concurrency address many of the latency issues I saw in
tests, generally by populating the cache.
* It is difficult to tune. On local storage the right setting depends on
the NVMe device concurrency. Using high numbers of read thread does not
help if query concurrency is also high.
* Each thread uses its own `Session` and there is no easy way to ensure
consistency across sessions.
* The thread pooling implementation was brittle in the face of session
exhaustion.

If this is really needed it shouldn't be awful to reimplement. If the
on-disk layout changes it might be better to
parallelize lookups for re-ranking and do a much smaller number of
reads, since raw vector bytes will usually
dominate other IO costs.