Implement naive i8 quantization (#43)
Implement a naive i8 quantization scheme based on the ScaNN
implementation.

This scheme does not train on a data set at all. Input float vectors are
l2 normalized before quantization, and we quantize
values assuming they may occupy any part of the range [-1,1]. The norm
is stored alongside the quantized vector and
query and doc norms are added to the distance when computing euclidean
distance.

Compared to raw float distances this has a loss of 3-3.5% recall but at
roughly a quarter of the cost.

There are a couple of potential extensions to this scheme:
* Train on data to bound the quantization range and increase precision.
* Quantize asymmetrically: score float query vectors against i8 doc
vectors. I tried this and it was worth about 1% recall precision.

The scoring function implementation is written in scalar code as
`simsimd` insists on performing cosine similarity
for i8 vectors for reasons that are unclear. This implementation
auto-vectorizes alright but in the long run might be
better off using `std::simd` or manually writing vectorized code for
this.