KMeans partitioning improvements (#45)
Add `iterative_balanced_kmeans` partitioning scheme.

This scheme limits itself to partitioning using an intermediate k value
with the goal of reaching the actual k value. At each step we may
partition centroids that are too large. A balancing factor is used to
decide when a partition is "too large" and must be split. This is much
faster than running k-means with a very large value (~1k) and yields a
more balanced distribution.

* Select random vectors for k-means initialization from the complete
dataset.
* Evaluate initial centers using a random subset of vectors to control
cost.
* Terminate based on max centroid distance. Summing did not work
particularly well when k was large.
* Add a tool to help observe iterative partitioning algorithm