Add a default simsimd feature and naive implementations if it is disabled (#49)
This is useful if you are struggling with a cc toolchain.

"simsimd" is a default feature, so default features must be disabled to
turn this off. When disabled distance
function implementations are replaced with trivial iterator-based
implementations. These are _probably_
pretty fast as it should be possible to auto-vectorize them, but success
is not guaranteed.