Overhaul balanced partition assignment for use by SPANN (#52)
With a 1M vector corpus the initial partition step would often fail to
converge if k was large (~128) but then would
perform overly granular clustering at subsequent steps if k was small
(~32 => ~250). Clustering in large chunks is
often slow due to the large number of vector comparisons that need to be
performed. This also often struggled to
partition the smallest chunks in situations where k is 2 or 3.

Revamp the interface to use a min and max centroid size instead of hard
constraining to a particular k value, but
continue to provide a maximum k value. Improve behavior around centroid
sub partitioning:
* If `k=2` use a binary partitioning scheme that terminates as soon as
it gets an acceptable split. There is future work here to keep trying a
bit longer in hopes of getting a better split.
* When sub partitioning produces under size partitions, iteratively
prune them out and reassign within the partition set. If only one sub
partition is viable, binary partition instead.

When partitioning ~1M 1536d vectors into centroids with size bounds
64..=192 it takes about 80 cpu seconds on
an M4 Macbook Air. AFAICT this seems to provide higher quality
partitioning than the existing code.