Implement a trivial quantizer (#53)
This encodes a float vector as a byte array, where each f32 is little
endian ordered.

This is useful for testing placement of full fidelity vectors in places
where we would otherwise use quantized vectors.
It is not a solution to the general problem of having two vector sets
but sometimes wanting to have just one.