Introduce a SPANN-like index implementation (#54)
This is an alternative implementation of
[SPANN](https://arxiv.org/pdf/2111.08566).

Aside from being implemented on top of WiredTiger, there are numerous
other differences:
* This leverages our iterative balanced kmeans implementations, which is
not quite the same as SPANNs balancing scheme.
* This uses means not mediods. As a result it duplicates the head index.
* The head index reuses our Vamana implementation instead of SPTAG.
* Defaults push you to build a head index of ~1% of the corpus size.

On a ~1M corpus with 256 centroids in search and binary quantized
postings I can get ~94.5% recall.
This is significantly slower than the Vamana implementation but very
little optimization has occurred.

Some smaller changes:
* `et` search harness can be built with serial searching to make
profiling possible.
* Print wt read block read stats.