SPANN build improvements (#57)
The biggest change here is allowing clustered means to be replaced with
the closest mediod, which greatly improves
recall at lower parameter settings. With means we would often get 0
secondary assignments for most vectors, but this
functions more naturally with mediods and we end up with an avg of R -
0.5 assignments for each vector. Based on
stats the new assignments also seem to be better balanced.

I can get ~93% recall with just 32 centroids in the search at about 5ms
latency with 4 replicas.

This differs from the SPANN paper in that we performing mediod
replacement at the end instead of incrementally at
each step.

Smaller changes:
* Upgrade rand and use a more specific, seeded RNG for clustering.
* Print assignment stats for primary, secondary, and total assignments.