Add support for table metadata, metadata cursors and config parsing (#61)
Using table metadata with json values will allow us to embed
configuration data in the metadata index
instead of relying on "special" keys in the table. Metadata cursors and
config parsing are necessary to
access this data in practice.

WT config strings are roughly compatible with JSON so we can continue to
use JSON metadata using
this new mechanism.