Unify vector representation in scoring around byte arrays (#62)
The observation here is that most vector scoring functions will use SIMD
instructions that allow unaligned loads,
so we can use `[u8]` as the rep instead of `[f32]` since we don't
_really_ have to worry about alignment, and
simplifies code for loading float vectors from WT which does not
guarantee alignment of returned pointers.

This opens a path to using the vamana code with a quantized
representation of the "raw" vectors.

Remove the simsimd feature, it's kind of moot since we won't be using
simsimd as heavily because simsimd accepts
slices and slices must be aligned. I think this is only a feature of the
rust API but none of the docs indicate anything
about alignment in the C API either.