Introduce a uniform scaled i8 quantizer (#66)
This quantizer uses the maximum absolute value across all dimensions to
scale each dimension. The scale and
l2 norm are stored along with the quantized vector to allow computing l2
and dot product distances in a way that
is compatible with f32 l2 and dot distances on the quantized
representation.

In a test where this is used for nav vectors, we can get 0.980560 recall
with re-ranking of the full candidate set and
0.97705 recall without -- a difference of 0.35%. This could replace f32
in the raw vector table if the infrastructure
supported it.

Like i8naive and unlike lucene's u7 scalar this quantizer does not
sample or measure the data set, meaning that changes
to the input vector set do not affect quantization.

Future work:
* A non-uniform scaled quantizer. In Matryoshka vector representations
some dimensions are more important than others and we could vary the
scaling factor across sets of dimensions to increase fidelity.
* Score f32 x quantized to improve recall. This requires significant
work to inject scorers in the query flow.
* Replace f32 vectors in the "raw" vector table to reduce the cost of
re-ranking.