Unify vector interfaces in graph handling (#72)
This makes it easier to treat nav and rerank vectors as interchangeable,
so that in the future may we go down to
just one vector table when high fidelity vectors are available.

Replace "raw" with "rerank" in most spots.

Use `app_metadata` to store configuration information instead of a
special key.