Packed `i4` scaled uniform quantization (#73)
This utilizes the same scaling/shaping behavior is `i8-scaled-uniform`
but quantizes to 4 bit values and packs two
dimension per byte, halving the amount of required storage. This comes
at a 3-4% recall hit.

During packing values are offset binary encoded to ease unpacking -- we
can mask and subtract to get sign extension
on negative values. This adds a couple of instructions on core decoding.