Add f16 vector format using the half crate (#79)
`f16` coding has near-zero quantization loss, and near zero distance
loss so long as values are converted to `f32`
before computing the distance.

All of the SIMD infrastructure for `f16` is not stabilized so
performance is quite poor, ~600ns to compute distance
across 1024 dimensions on an M4 Mac, ~4-5x the cost of doing so on f32
vectors. The compiler generated code
uses inline assembly without loop unrolling and absolutely no SIMD for
this. This is fixable by using C, I just don't feel
like doing that right now.

Refactor so that creating the coder accepts a similarity and in the case
of f32 and f16 codings, chooses to normalize
the vector during encoding. This replaces the special raw-l2-norm format
and a bunch of adjustments around that.