Pass `Cow` when creating `QueryVectorDistance` (#80)
This allows the caller to pass an owned buffer instead of needing to
maintain a reference.

This makes it simpler to handle cases where we might want to use
different query representation, which was
being encoded in the type name as a lifetime. Taking ownership allows a
single lifetime (`'static`).