Add an `i16` scaled uniform quantization scheme (#82)
This is like `i8` but scaled to 16 bits, with much lower quantization
and distance loss than `i8`. It should behave at least
as well as `f16` according to loss metrics.

This is faster than `i4` and `i8` to score against `f32` vectors,
probably mostly owing to fewer widening instructions, but
is not as fast to perform fully quantized scoring -- fewer elements per
vector and no `SDOT` path.

This could be the basis for a split representation codec -- metadata and
the high byte are used for navigation, but can
be merged with the low byte for higher precision when re-ranking.