Add a tool to compute recall against a ground truth set (#83)
This is a better grounded version of `distance-loss`.

This leverages a pre-computed topN set so we can skip `f32` scoring.
Document vectors are quantized and then
compared against all the query vectors, which are then grouped by query.
This is done using shared state because
managing this inversion directly with `rayon` leads to a ton of
allocations.

One thing I've learned here is that the scaled-uniform quantization
functions are _very_ slow.