Add `VectorSimilarity::Cosine` (#86)
This allows us to run angular distance in two modes:
* `Dot` for l2 normalized inputs
* `Cosine` for unnormalized inputs

Making this distinction allows us to avoid dot product/normalization in
vector coding and distance computation
in spots if we know that the vectors are normalized already.

Remove asymmetric binary coding. It doesn't make much sense at the
moment and should be a QueryVectorDistance
instead of a full-fledged codec. We'd also likely be better off if this
were replaced with an i8 scaled-uniform based
query encoding and decoding.