Add truncatedf32 vector coding (#88)
Allow truncating at a fixed dimensionality. This is useful for MRL model
output where certain ranges of dimensions
are "more important" than others to the final similarity result and a
truncated vector can be useful on its own.

There is some temptation to generalize this to allow truncating and then
applying an arbitrary vector coding. In practice
this is a huge message as `F32VectorCoding` must become `!Copy`.