Implement a higher fidelity quantized query vector distance for binary vectors (#90)
This uses i8-scaled-uniform coding for the high fidelity component of
the vector comparison. This allows us to
table decode the binary vector into a form that is equivalent to a unit
vector in the same encoding as the i8 scaled
vector query relatively cheaply. This is worth about 7% recall or NDCG
improvement of quantized comparison, but is
also 7-8x slower than plain hamming distance.