Improve LVQ2 accuracy when very few primary bits are used (#97)
In this case interval optimization may pull in the bounds so far that
the delta value is not sufficient to accurately
encode residual values. This is particularly acute when 1 bit is used
for the primary vector.

To fix this explicitly pick an interval for the residual vector that is
encoded alongside the vector contents. We do
not do anything clever here, choosing the maximum value of the primary
vector delta or the gaps between the
(minimum, primary lower) or (maximum, primary upper). This ensures that
the residual vector can encode the
minimum and maximum values in the original vector, substantially
improving accuracy.

Fix a bug in the aarch64 accelerated implementation where we were
relying on the value saturating when converting
from f32 to u32 even though that is not the case.

With this change `lvq2x1x8` is more accurate than `lvq1x8`, which is
expected given that it uses 1 extra bit.