Add 12 and 16 bit residuals to lvq2 (#99)
lvq2 is intended to be stored with split representation where the
residual is only used for reranking, so allow
more residual bits to improve accuracy.

In quantization-recall tests lvq2x1x12 does slightly better than f16 in
exhaustive recall situations while using 82%
of the space and only just ~7% of the memory space.