Improve lvq dot performance on aarch64 (#105)
This path was not manually optimized on aarch64. The 8-bit
implementation has the same performance; the 1 bit and
4 bit are much faster. 4 bit benefits in particular because the packing
arrangement made it difficult to optimize, but we
know that the order of computation is not important so long as
dimensions are paired correctly.

TIL the compiler is generating ldp (32 byte load) and performing dot
independently in two registers for 8 bit dot, and
this optimization can be extended into the 4 bit implementation.

```
lvq1x1/doc/dot          time:   [6.8411 ns 6.8684 ns 6.9008 ns]
                        change: [−38.070% −37.715% −37.379%] (p = 0.00 < 0.05)
lvq1x4/doc/dot          time:   [21.231 ns 21.269 ns 21.310 ns]
                        change: [−57.553% −57.276% −57.075%] (p = 0.00 < 0.05)
lvq1x8/doc/dot          time:   [27.817 ns 27.850 ns 27.890 ns]
                        change: [−0.2463% −0.0105% +0.2355%] (p = 0.93 > 0.05)
```