Implement SOAR distance computation for centroid replica selection (#108)
SOAR is used to choose centroid assignments after the primary (closest)
assignment.

In SPANN we are forced to use mediods with the existing RNG pruning
because it will fail to assign replicas in most
cases due to occlusion. SOAR works well in this case and the resulting
clusters are more balanced.

SOAR distance computation only works on float vectors, so improve
decoding support in vector coders.