Improve WiredTiger Transaction APIs (#110)
Add a `TransactionGuard` struct as a transaction scope.

Users of this API may explicitly `commit()` or `rollback()` theirs
changes, consuming the scope guard. If no action is
taken the transaction is rolled back at the end of the scope. This API
has some caveats:
* Existing transaction methods are still exposed so users can break the
invariants. Don't to that. I will eventually fix this.
* This API does not prevent attempts to nest transactions at compile
time. Fixing this would be complicated and I'm not sure it would make
this any easier to use.

Implement `Iterator` for `TypedCursorGuard` to make get/open cursor API
equivalent.