Implement wavelet scoring in SPANN (#112)
Wavelet scoring is implement with a "fast" scorer that can be scored in
quantized space and a "slow" scorer that
typically directly compares an f32 vector against the quantized
representation. We then maintain a result queue for
each scorer but only compute+insert in the "slow" queue if the result is
successfully pushed into the "fast" queue.
This is especially effective with two-level LVQ as the "fast" scorer can
produce a primary quantized query vector and
compare that to the stored primary vector, then use the second level
residuals for the "slow" scorer.

With this arrangement `lvq2x1x8` is *3x* faster than `lvq1x8` when
searching with the same parameters with
negligible (< 0.1%) recall cost. I can score 3x more centroids in the
same envelope and improve recall by 4.2% or
score 2x more for a 3% recall improvement and bank some capacity. It
might also be possible to make this faster
by splitting the lvq representation and packing it as <all
primaries><all residuals> within a tail index.

SPANN is seek bound but not necessarily IO bound, so the costs for this
are probably manageable at scale.