Make wt key/value formatting faster (#113)
Refactor key/value formatting for cursors so that it can be more
performant.

In general these APIs are not very fast when you can't inline them, and
I was unable to get LTO to inline them.
The streaming API performs an allocation for every key/value read
(except trivial formats in the current framework);
the one shot pack/unpack APIs are more efficient but it's still not
great for compound keys and was showing up in
profiles for SPANN searches.

Refactor the API so that we use the one-shot pack/unpack APIs under the
hood and provide implementations for
packing/unpacking up to 3 values. The new API also allow more latitude
w.r.t. the format so for SPANN keys we are
trivially encoding each key as a 12-byte value without using the
underlying WT APIs which is significantly faster.

This does regress in one area: for trivial formats ("u") it is no longer
possible to "encode" without copying the data
into another buffer. This is unlikely to be a big issue in practice and
it may be possible to fix with a similar specialized
API to the one that existed before.