Introduce a centroid selection based on target vector scoring count (#119)
Introduce an abstraction that can either select a fixed number of
centroids (like today) or use `CentroidStats` to
select centroids until we reach a fixed count.

This frequently results in scoring fewer vectors while still getting
consistent recall. It may be less necessary if we
were better at balancing the contents of the centroids.

The observation is that for our set of ~2M voyage dbpedia vectors we can
score ~25k and get > 90% of recall for
any value of `k` so long as `--posting-candidates` is ~10x of k.