Fix in-place delete handling in vamana (#123)
Borrow parts of Algorithm 6 from
https://www.vldb.org/pvldb/vol18/p5166-upreti.pdf to improve in-place
deletes.

Much of this algorithm is not relevant because our graph is undirected
so FoF traversal to try to discover most of
the in-links is unnecessary. We do still compute distances between all
the edges on the deleted vertex and select
the top c edges (in our case 50% of max_edges) to integrate back into
the outbound edges from each vertex. This
is complicated in practice because of the symmetric/undirected
invariant.

The previous scheme attempted to insert edges when there was room so it
would never trigger any sort of pruning
and did nothing to ensure any sort of diversity.