Expose vamana alpha hyper parameter (#125)
Expose the alpha parameter, which extends the length of edges that may
be kept while performing relative neighbor
graph pruning. This is exposed everywhere we may create a vamana index.
Evidence suggests I may need to use this
on the SPANN head index to keep more edges on a heavily mutated graph.

Fix a bug in the existing implementation where we were comparing the
wrong distances -- it should be the distance
from the selected edges to the vertex rather than the candidate edge.

In manual tests this greatly increases the number of edges kept,
particularly when max_alpha is increased.

Remove `GraphLayout` because it is unused.