Remove graph clustering code (#126)
This was an experiment to see if we could speed graph insertion by
inserting "like" vectors at roughly the same time,
inspired by lucene bp vector reorderer. This might be necessary but it
is not sufficient -- the vectors must also be
located very close together in storage/ram to get any benefit.