Accelerate vector decoding for f16 and lvq (#130)
This gets invoked a lot more often in SPANN where we need to decode
vectors to re-cluster them.
It might also have some purchase in the bulk load code path where we
might be better off clustering
a quantized representation of the vectors to reduce resource
utilization.

While this does require specific code paths there's a lot of overlap
with f32 x quantized scoring paths so
there is some internal re-use of code.