Replace SPANN k-means with one that performs better on larger than memory data sets (#131)
This implementation maintains a buffer of vectors. For any given subset:
* If the subset is larger than the buffer, sample to buffer size and
perform k-means on the sample.
* Otherwise perform k-means on a subset in the buffer.

This reduces IO compared to batch k-means where each iteration reads
random vectors. The bet is that with a
large enough sample it can represent that dataset to perform fairly
coarse partitioning.

Use `Advice::Random` for the input dataset. This _greatly_ reduces IO on
linux and makes posting list upload
tolerable without resorting to a merge sort.