Parallelize hierarchical kmeans (#132)
After initial clustering take the set of generated partitions and split
them into N pieces, then assign threads to perform
hierarchical clustering on these groups.

The existing single-threaded flow may oscillate between being IO bound
and being CPU as it alternates between reading
vectors to sample (or buffer the entire centroid) and performing k-means
computations. This could be parallelized
by the chosen k-value but because the buffer is so large (1GB by
default) we want to control the level of parallelism.

Process blocks of posting keys in parallel when creating the posting
index. This parallelizes the random IO that occurs
when reading vectors as well as posting vector encoding which is
expensive (single digit usecs/vector).