Implement patience algorithm for graph search early termination (#134)
The idea is to examine how many results change in each round of visited
vertex(es) and if the change is below a
threshold for N consecutive rounds then terminate early.

It isn't obvious how to use this most effectively when candidate count
is relatively small -- the count parameter is
very difficult to tune and recall can be quite poor. When candidate
count is large (thousands) this can quite effectively
reduce the amount of work done with minimal recall loss. It's also quite
effective with heavy quantization where we
struggle to distinguish between candidates based on scores and reranking
will cover many losses.

See
https://cs.uwaterloo.ca/~jimmylin/publications/978-3-031-88714-7_39.pdf