Improve binary partitioning convergence story (#135)
Binary partitioning would fail to converge a handful of times per 10k
clusters. There were a few issues:
* If inertia got worse it wouldn't fall back correctly, even if the
*starting* centroids had an acceptable split. This was impossible to fix
in the current structure because we didn't know enough about previous
centroid candidates.
* If initial centroid selection was poor it might not be possible to
converge. Combined with the 50/50 split we might get "stuck" in a
particular spot and compute hundreds of noop iterations.

Refactor into a candidate struct and iter state for re-used buffers. We
retain old candidates so we can "go back" if the
next iteration is worse. Use an iteration amplified adjustment for the
split so that we "try harder" to change the assignments
if we get stuck and the split is not acceptable.

Use k-means++ style initialization of the centroids and perform multiple
initializations, taking the best one.