Add timestamp support to wt_mdb (#136)
This allows committing at and reading at timestamps within transactions.

This refactors some of the APIs around transactions to not only add
timestamp related fields but to improve the
ergonomics of options inside and outside of the crate.

Using timestamps would allow us to perform reads in multiple threads as
part of a single query so long as all commits
are timestamped. The "main" query thread would pick a timestamp to read
from and would communicate this to other
read threads. This would work with all bulk loads but continuous
workloads will need to integrate write timestamps.