Implement SPANN rebalancing routines (#138)
These routines merge small centroids and split oversize centroids in a
way that tries to maintain the property
that each vector is closest to its assigned primary centroid.

Includes additional infrastructure around updating centroid assignments
and related stats tables. For now this only
supports a single replica as it is more complicated to implement for
multiple replicas.

The command line tool is extended to run rebalancing iteratively until
convergence.